+++ HexGame/src/hexagon.h
@@ -9,8 +9,8 @@ class Hexagon : public QObject
     Q_PROPERTY(QString color READ getColorStr NOTIFY colorChanged)
 
 public:
-    Hexagon(int id, QObject* parent = nullptr): QObject(parent), m_color(Game::Color::Empty), id(id) {};
-    const int id;
+    Hexagon(uint8_t id, QObject* parent = nullptr): QObject(parent), m_color(Game::Color::Empty), id(id) {};
+    const uint8_t id;
     QString getColorStr();
     Game::Color getColor();
     void setColor(Game::Color color);
